requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.4
python-multipart==0.0.18
selenium==4.27.1
webdriver-manager==4.0.2
apscheduler==3.10.4
sqlalchemy==2.0.36
pyperclip==1.9.0
python-dotenv==1.0.1
Pillow==11.1.0
